requirements.txt
transformers>=4.30.0
sentence-transformers>=2.2.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
scikit-learn>=1.2.0
pymediainfo>=6.0.0
Pillow>=10.0.0
numpy>=1.24.0
ffmpeg-python>=0.2.0 